♻️ :: refactoring main page
