+++ app/src/main/java/com/msg/gcms/presentation/view/main/MainActivity.kt
@@ -1,13 +1,18 @@
 package com.msg.gcms.presentation.view.main
 
+import android.util.Log
 import android.view.View
 import androidx.activity.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
+import coil.load
 import com.msg.gcms.R
 import com.msg.gcms.databinding.ActivityMainBinding
 import com.msg.gcms.presentation.base.BaseActivity
+import com.msg.gcms.presentation.utils.enterActivity
 import com.msg.gcms.presentation.view.club.detail.DetailFragment
+import com.msg.gcms.presentation.view.clubmaker.MakeClubActivity
+import com.msg.gcms.presentation.view.profile.ProfileActivity
 import com.msg.gcms.presentation.viewmodel.ClubDetailViewModel
 import com.msg.gcms.presentation.viewmodel.MainViewModel
 import dagger.hilt.android.AndroidEntryPoint
@@ -22,12 +27,16 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     override fun viewSetting() {
         initBottomNav()
         getProfileImage()
+        clickProfile()
+        clickMakeClubBtn()
     }
 
     override fun observeEvent() {
         observeBottomNav()
         observeSetNav()
         getProfileImage()
+        observeProfileImage()
+        observeClubName()
         observeProfileImage()
     }
 
@@ -37,13 +46,40 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
 
     private fun observeProfileImage() {
         mainViewModel.getProfile.observe(this) {
-            mainViewModel.getProfile()
+            Log.d("TAG", "observeProfileImage: $it")
+            if (it != null) binding.profileBtn.load(mainViewModel.getProfile())
+            else binding.profileBtn.setImageResource(R.drawable.ic_profile)
+        }
+    }
+
+    private fun clickProfile() {
+        binding.profileBtn.setOnClickListener {
+            enterActivity(this, ProfileActivity())
+        }
+    }
+
+    private fun clickMakeClubBtn() {
+        binding.addClubBtn.setOnClickListener {
+            enterActivity(this, MakeClubActivity())
+        }
+    }
+
+    private fun observeClubName() {
+        mainViewModel.clubName.observe(this) {
+            if (binding.clubNameTxt.text != mainViewModel.clubName.value) {
+                mainViewModel.getClubList()
+            }
+            binding.clubNameTxt.text = mainViewModel.clubName.value
         }
     }
 
     override fun onResume() {
         super.onResume()
-        if (intent.getBooleanExtra("isProfile", false) && detailViewModel.isProfile.value == false) {
+        if (intent.getBooleanExtra(
+                "isProfile",
+                false
+            ) && detailViewModel.isProfile.value == false
+        ) {
             detailViewModel.getDetail(intent.getLongExtra("clubId", 0))
             // detailViewModel.setResult(intent.getSerializableExtra("result") as ClubDetailData)
             detailViewModel.setIsProfile(true)
